# Move Moshi to extension for 17.5% load time reduction